add protected resolve method to encapsulate response cleanup
