+++ src/IApi.ts
@@ -13,4 +13,6 @@ export default interface IApi {
   put: FetchMethod;
   patch: FetchMethod;
   delete: FetchMethod;
+ 
+  fetch: (url: string, options: RequestInit) => Promise<any>;
 }
